// src/types/APIPegadaianTypes.ts
export interface IKTPPassion {
  cif: string;
  hp: string;
  ibuKandung: string;
  jenisKelamin: string;
  kodeCabang: string;
  namaNasabah: string;
  noIdentitas: string;
  telp: string;
  tglLahir: string;
  tglKyc: string;
  statusKyc: string;
  statusKawin: string;
  kewarganegaraan: string;
  tipeIdentitas: string;
  idKelurahan: string;
  jalan: string;
  statusPegawai: boolean;
}